Updated Create page script names for clarity.
