+++ Inspicio/Views/Images/Create.cshtml
@@ -225,4 +225,4 @@
     });
 
 </script>
-<script type="text/javascript" src="~/dist/multipleImages-New.bundle.js"></script>
+<script type="text/javascript" src="~/dist/multipleImages.bundle.js"></script>
