+++ Inspicio/wwwroot/js/createDropzone-old.js
